+++ src/fields/Matrix.php
@@ -212,49 +212,55 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
     /**
      * Returns all of the block types' fields.
      *
+     * @param int[]|null $typeIds The Matrix block type IDs to return fields for.
+     * If null, all block type fields will be returned.
      * @return FieldInterface[]
      */
-    public function getBlockTypeFields(): array
+    public function getBlockTypeFields(?array $typeIds = null): array
     {
-        if (isset($this->_blockTypeFields)) {
-            return $this->_blockTypeFields;
-        }
+        if (!isset($this->_blockTypeFields)) {
+            $this->_blockTypeFields = [];
 
-        if (empty($blockTypes = $this->getBlockTypes())) {
-            return $this->_blockTypeFields = [];
-        }
+            if (!empty($blockTypes = $this->getBlockTypes())) {
+                // Get the fields & layout IDs
+                $contexts = [];
+                $layoutIds = [];
+                foreach ($blockTypes as $blockType) {
+                    $contexts[] = 'matrixBlockType:' . $blockType->uid;
+                    $layoutIds[] = $blockType->fieldLayoutId;
+                }
 
-        // Get the fields & layout IDs
-        $contexts = [];
-        $layoutIds = [];
-        foreach ($blockTypes as $blockType) {
-            $contexts[] = 'matrixBlockType:' . $blockType->uid;
-            $layoutIds[] = $blockType->fieldLayoutId;
-        }
+                /** @var FieldInterface[] $fieldsById */
+                $fieldsById = ArrayHelper::index(Craft::$app->getFields()->getAllFields($contexts), 'id');
 
-        /** @var FieldInterface[] $fieldsById */
-        $fieldsById = ArrayHelper::index(Craft::$app->getFields()->getAllFields($contexts), 'id');
+                // Get all the field IDs grouped by layout ID
+                $fieldIdsByLayoutId = Craft::$app->getFields()->getFieldIdsByLayoutIds($layoutIds);
 
-        // Get all the field IDs grouped by layout ID
-        $fieldIdsByLayoutId = Craft::$app->getFields()->getFieldIdsByLayoutIds($layoutIds);
+                // Assemble the fields
+                foreach ($blockTypes as $blockType) {
+                    if (isset($fieldIdsByLayoutId[$blockType->fieldLayoutId])) {
+                        $fieldColumnPrefix = 'field_' . $blockType->handle . '_';
 
-        // Assemble the fields
-        $this->_blockTypeFields = [];
-
-        foreach ($blockTypes as $blockType) {
-            if (isset($fieldIdsByLayoutId[$blockType->fieldLayoutId])) {
-                $fieldColumnPrefix = 'field_' . $blockType->handle . '_';
-
-                foreach ($fieldIdsByLayoutId[$blockType->fieldLayoutId] as $fieldId) {
-                    if (isset($fieldsById[$fieldId])) {
-                        $fieldsById[$fieldId]->columnPrefix = $fieldColumnPrefix;
-                        $this->_blockTypeFields[] = $fieldsById[$fieldId];
+                        foreach ($fieldIdsByLayoutId[$blockType->fieldLayoutId] as $fieldId) {
+                            if (isset($fieldsById[$fieldId])) {
+                                $fieldsById[$fieldId]->columnPrefix = $fieldColumnPrefix;
+                                $this->_blockTypeFields[$blockType->id][] = $fieldsById[$fieldId];
+                            }
+                        }
                     }
                 }
             }
         }
 
-        return $this->_blockTypeFields;
+        $fields = [];
+
+        foreach ($this->_blockTypeFields as $blockTypeId => $blockTypeFields) {
+            if ($typeIds === null || in_array($blockTypeId, $typeIds)) {
+                array_push($fields, ...$blockTypeFields);
+            }
+        }
+
+        return $fields;
     }
 
     /**