+++ src/services/Elements.php
@@ -1226,6 +1226,8 @@ class Elements extends Component
             static::$duplicatedElementIds[$element->id] = $mainClone->id;
             static::$duplicatedElementSourceIds[$mainClone->id] = $element->id;
 
+            $mainClone->newSiteIds = [];
+
             // Propagate it
             foreach ($supportedSites as $siteInfo) {
                 if ($siteInfo['siteId'] != $mainClone->siteId) {
@@ -1303,6 +1305,10 @@ class Elements extends Component
 
                     if (!$this->_saveElementInternal($siteClone, false, false)) {
                         throw new InvalidElementException($siteClone, "Element $element->id could not be duplicated for site {$siteInfo['siteId']}: " . implode(', ', $siteClone->getFirstErrors()));
+                    }
+
+                    if ($siteClone->isNewForSite) {
+                        $mainClone->newSiteIds[] = $siteClone->siteId;
                     }
                 }
             }
@@ -2618,7 +2624,7 @@ class Elements extends Component
                 ]);
             }
 
-            if ($isNewSiteElement = empty($siteSettingsRecord)) {
+            if ($element->isNewForSite = empty($siteSettingsRecord)) {
                 // First time we've saved the element for this site
                 $siteSettingsRecord = new Element_SiteSettingsRecord();
                 $siteSettingsRecord->elementId = $element->id;
@@ -2635,7 +2641,7 @@ class Elements extends Component
             }
 
             // Update our list of dirty attributes
-            if ($trackChanges && !$isNewSiteElement) {
+            if ($trackChanges && !$element->isNewForSite) {
                 array_push($dirtyAttributes, ...array_keys($siteSettingsRecord->getDirtyAttributes([
                     'slug',
                     'uri',