+++ src/services/Revisions.php
@@ -20,8 +20,8 @@ use craft\helpers\Queue;
 use craft\queue\jobs\PruneRevisions;
 use Throwable;
 use yii\base\Component;
+use yii\base\Exception;
 use yii\base\InvalidArgumentException;
-use yii\db\Exception;
 
 /**
  * Revisions service.
@@ -74,7 +74,7 @@ class Revisions extends Component
 
         $lockKey = 'revision:' . $canonical->id;
         $mutex = Craft::$app->getMutex();
-        if (!$mutex->acquire($lockKey)) {
+        if (!$mutex->acquire($lockKey, 3)) {
             throw new Exception('Could not acquire a lock to save a revision for element ' . $canonical->id);
         }